⚰️ (cleaning) removed some dead code
